Created edit_data and updated query to details pg
edit_data will eventually allow a user to change the data for some
movie. Query action has been updated in index.js and results.js so that
they can pass the actors and role names along to the details page, but
for some reason right now the result of the nested query is not passing
on (it's just saying the result parameter of the callback is an object).